Add default value for edited_at column
Without this, a clean clone doesn't migrate on VERSION=8.
I'm not sure this is the proper solution, but it seems ok,
especially since the next command sets all of these values
equal to updated_at.

I wonder if there's a better way to do this, but I don't
understand migrations as well as I should.