Don't throw exception when viewing empty tag.
I'm settings my posts date to future as a replacement for the missing "draft" feature. If you do so the post will not be displayed in posts index or posts index by tag, but the tag will be visible in the navigation bar. Clicking a new for which the article isn't published yet causes an exception.

This commits makes it display "There are no posts yet".

The real fix to this problem would be filtering out tags with only unpublished posts, but this would either require a lot more coding or would slow down tag list terribly.