Load jquery unobtrusive js driver.
At least one ujs driver needs to be loaded if enki is to be used with Rails 3. Since Prototype is not included in any layout I'm adding Jquery that is used everywhere in Enki.

Without this driver things like the Logout button in Admin panel did not work. It was pointing to GET admin_session_url instead of
DELETE admin_session_url.

The driver required jquery update.