Fix code formatting by restricting coderay and lesstile versions
The Gemfile allowed me to install incompatible verisons of lesstile and
coderay (1.0.1 and 1.0.5, respectively).  The result was this exception
when formatting a post:

     NoMethodError:
       undefined method `div' for #<String:0x00000103b7daf0>

The incompatibility is fixed in lesstile 1.1.0.

Request the fixed version in the Gemfile, and pair it with a
version of coderay that is known to be compatible.  Add an integration
spec to ensure future compatability.