Track db/schema.rb in version control
Rails best practice is to track db/schema.rb in version control, but
Enki does not track it.  Generally, version control best practices
dictate that generated files are not tracked, and db/schema.rb is
a generated file in some sense.  But the default Rails .gitignore does
not exclude it because:

    Migrations, mighty as they may be, are not the authoritative source
    for your database schema. That role falls to either db/schema.rb or
    an SQL file which Active Record generates...
    http://guides.rubyonrails.org/migrations.html#what-are-schema-files-for

Remove db/schema.rb from .gitignore and add it to Git version control.
This version of db/schema.rb was generated by running the current set of
migrations (which also regenerates db/schema.rb).