Add body_filter to Post to allow authoring in HTML
I have some old posts that I authored in HTML, and I anticipate
continuing to do so on occasion.  At a minimum, I need to be able to
preserve the markup of the posts I import into Enki Blog.  Ideally, I'd
also be able to easily use HTML to author posts in the future.
(Embedding tags in Textile is not enough.)

Add a body_filter attribute to the Post model.  Adjust apply_filter to
be able to handle the default textile-to-XHTML converstion or a null
conversion that assumes the post was authored as HTML already.