Whitelisting parameters required for successful display of posts.
The move to Rails 4 and strong parameters necessitates whitelisting of
parameters being processed by the controller. The exclusion of the
:tag_list, :published_at_natural and :slug parameters from the
whitelist means that these data are not being saved to the db and hence
no new post will be displayed on the front end, no new tags will be
saved and the post slug cannot be updated after initial creation. This
commit fixes these issues.