Applied fix for logging in via OpenID to the admin area.
Rails now contains code that resets the session if the CSRF request
forgery check fails. This was happening on the OpenID callback when
logging in to the Enki admin area, which broke OpenID login.

This commit disables the CSRF check only for OpenID callbacks when
logging in to the Enki admin area.

Fixes #91.