Fixed incorrect tag name behaviour.
When a tag name contains a dot character, Rails interprets this as the
separator for a formatted route and splits the tag name into the section
before the dot which is interpreted as the expected :tag parameter, and
the section after the dot which is interpreted as an unexpected :format
parameter.

This commit restores correct behaviour in the event that a tag name
contains a dot character and includes spec coverage.

Fixes #93.